pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "Claude-powered crypto auto-trading system"
requires-python = ">=3.12,<3.14"
dependencies = [
    "ccxt>=4.4,<5",
    "websockets>=13,<14",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "pyyaml>=6,<7",
    "sqlalchemy[asyncio]>=2.0,<3",
    "aiosqlite>=0.20",
    "redis[hiredis]>=5.2,<6",
    "anthropic>=0.45,<1",
    "httpx>=0.28,<1",
    "numpy>=2.1,<3",
    "structlog>=24,<26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6",
    "ruff>=0.8",
    "mypy>=1.13",
]
postgres = [
    "asyncpg>=0.30",
    "psycopg[binary]>=3.2",
]
monitoring = [
    "slack-sdk>=3.33",
    "python-telegram-bot>=21",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
